# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BHRM Teams API
  version: 2.0.0
  description: |
    Enterprise Task-Centric Execution Platform — REST API.

    ## Authentication
    All protected routes require `Authorization: Bearer <token>` obtained from
    `POST /auth/login` or `POST /auth/refresh`.

    ## Response envelopes
    | Type | Shape |
    |---|---|
    | Single resource | `{ "success": true, "data": { … } }` |
    | Collection | `{ "success": true, "data": [ … ], "meta": { total, page, limit, totalPages } }` |
    | Error | `{ "success": false, "error": { "code": "…", "message": "…", "statusCode": 4xx } }` |

    ## Role model
    | Role | Description |
    |---|---|
    | `admin` | Company owner — full access to all company resources |
    | `staff` | Invited member — access scoped to assigned tasks |

    ## Endpoint naming conventions
    - Resource-based URLs, nouns only
    - State transitions via `PATCH /{resource}/{id}/status`
    - Bulk assignment via `PATCH /{resource}/{id}/assignees`
    - Avoid custom action verbs in URLs
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [admin, staff]
...
    SprintStatus:
      type: string
      enum: [draft, active, completed]
    TaskStatus:
      type: string
      enum: [todo, in_progress, review, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, critical]
...
    PaginatedMeta:
      type: object
      properties:
        total:      { type: integer, example: 42 }
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        totalPages: { type: integer, example: 3 }

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code:       { type: string, example: NOT_FOUND }
            message:    { type: string, example: Resource not found }
            statusCode: { type: integer, example: 404 }

    # ── User ─────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        fullName:  { type: string, example: Jane Smith }
        email:     { type: string, format: email }
        avatarUrl: { type: string, format: uri, nullable: true }
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          properties:
            role:            { $ref: '#/components/schemas/UserRole' }
            companyId:       { type: string, format: uuid, nullable: true }
            isEmailVerified: { type: boolean }
            lastActiveAt:    { type: string, format: date-time, nullable: true }
            createdAt:       { type: string, format: date-time }
            updatedAt:       { type: string, format: date-time }

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    StepProgress:
      type: object
      properties:
        total:      { type: integer }
        completed:  { type: integer }
        percentage: { type: integer, minimum: 0, maximum: 100 }
...
    Task:
      type: object
      properties:
        id:                { type: string, format: uuid }
        sprintId:          { type: string, format: uuid }
        parentTaskId:      { type: string, format: uuid, nullable: true }
        name:              { type: string }
        description:       { type: string, nullable: true }
        status:            { $ref: '#/components/schemas/TaskStatus' }
        priority:          { $ref: '#/components/schemas/TaskPriority' }
        startDate:         { type: string, format: date, nullable: true }
        plannedDueDate:    { type: string, format: date, nullable: true }
        actualDueDate:     { type: string, format: date, nullable: true }
        plannedEffortPh:   { type: number }
        estimatedEffortPh: { type: number }
        actualEffortPh:    { type: number }
        slippagePh:
          type: number
          description: 'actualEffortPh − plannedEffortPh; computed on read, never stored'
        isOverdue:         { type: boolean }
        stepProgress:      { $ref: '#/components/schemas/StepProgress' }
        owner:             { $ref: '#/components/schemas/UserSummary' }
        assignees:
          type: array
          items: { $ref: '#/components/schemas/UserSummary' }
        subTaskCount: { type: integer }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            sprint:
              nullable: true
              type: object
              properties:
                id:        { type: string, format: uuid }
                projectId: { type: string, format: uuid }
                name:      { type: string }
                goal:      { type: string, nullable: true }
                status:    { $ref: '#/components/schemas/SprintStatus' }
                startDate: { type: string, format: date, nullable: true }
                endDate:   { type: string, format: date, nullable: true }
            steps:
              type: array
              items: { $ref: '#/components/schemas/TaskStep' }
            recentComments:
              type: array
              items: { $ref: '#/components/schemas/Comment' }
            recentAttachments:
              type: array
              items: { $ref: '#/components/schemas/Attachment' }
...
    TaskAssigneeHistoryEntry:
      type: object
      properties:
        user:         { $ref: '#/components/schemas/UserSummary' }
        assignedBy:   { $ref: '#/components/schemas/UserSummary' }
        assignedAt:   { type: string, format: date-time }
        unassignedAt: { type: string, format: date-time, nullable: true }
        isActive:     { type: boolean }

    # ── Step ──────────────────────────────────────────────────────────────────
    TaskStep:
      type: object
      properties:
        id:        { type: string, format: uuid }
        taskId:    { type: string, format: uuid }
        title:     { type: string }
        isChecked: { type: boolean }
        order:     { type: integer }
        checkedAt: { type: string, format: date-time, nullable: true }
        checkedBy: { $ref: '#/components/schemas/UserSummary', nullable: true }
        createdAt: { type: string, format: date-time }

    # ── Attachment ────────────────────────────────────────────────────────────
    Attachment:
      type: object
      properties:
        id:          { type: string, format: uuid }
        taskId:      { type: string, format: uuid }
        fileName:    { type: string, example: design-spec.pdf }
        fileSize:    { type: integer, description: 'Bytes; max 10 485 760 (10 MB)' }
        mimeType:    { type: string, example: application/pdf }
        description: { type: string, nullable: true }
        uploader:    { $ref: '#/components/schemas/UserSummary' }
        downloadUrl: { type: string, format: uri }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }

    # ── Note ─────────────────────────────────────────────────────────────────
...
    CommentReply:
      type: object
      properties:
        id:        { type: string, format: uuid }
        content:   { type: string }
        author:    { $ref: '#/components/schemas/UserSummary', nullable: true }
        createdAt: { type: string, format: date-time }
    CommentReaction:
      type: object
      properties:
        id:    { type: string, format: uuid }
        emoji: { type: string, example: '👍' }
        user:  { $ref: '#/components/schemas/UserSummary' }
    CommentMention:
      type: object
      properties:
        userId:   { type: string, format: uuid }
        fullName: { type: string }
        email:    { type: string, format: email }
    Comment:
      type: object
      properties:
        id:      { type: string, format: uuid }
        taskId:  { type: string, format: uuid }
        content:
          type: string
          description: 'Content of the comment; "[deleted]" if soft-deleted'
        author:
          $ref: '#/components/schemas/UserSummary'
          nullable: true
        mentions:
          type: array
          items: { $ref: '#/components/schemas/CommentMention' }
        replies:
          type: array
          items: { $ref: '#/components/schemas/CommentReply' }
        reactions:
          type: array
          items: { $ref: '#/components/schemas/CommentReaction' }
        isEdited:  { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    # ── Activity ──────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role or resource ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    MessageOk:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean }
              data:
                type: object
                properties:
                  message: { type: string }

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
...
    sprintId:     { name: sprintId,     in: path,  required: true, schema: { type: string, format: uuid } }
    taskId:       { name: taskId,       in: path,  required: true, schema: { type: string, format: uuid } }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    search:
      name: search
      in: query
      schema: { type: string }

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    patch:
      tags: [Users]
      summary: Update current user profile (fullName)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName: { type: string }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /users/me/password:
    patch:
      tags: [Users]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword:     { type: string, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/tasks:
    get:
      tags: [Users]
      summary: List tasks assigned to the current user
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/TaskStatus' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/TaskPriority' }
      responses:
        '200':
          description: Paginated task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /sprints/{sprintId}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in a sprint (admin sees all; staff sees assigned only)
      parameters:
        - $ref: '#/components/parameters/sprintId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/TaskStatus' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/TaskPriority' }
        - $ref: '#/components/parameters/search'
        - name: parentTaskId
          in: query
          description: '"null" (string) returns top-level tasks only; omit to return all'
          schema: { type: string }
        - name: assigneeId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paginated task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      tags: [Tasks]
      summary: Create a task in a sprint (admin only; blocked for completed sprints)
      parameters:
        - $ref: '#/components/parameters/sprintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:              { type: string }
                description:       { type: string }
                priority:          { $ref: '#/components/schemas/TaskPriority' }
                parentTaskId:
                  type: string
                  format: uuid
                  description: 'Must belong to same sprint; supports unlimited nesting'
                startDate:         { type: string, format: date }
                plannedDueDate:    { type: string, format: date }
                plannedEffortPh:   { type: number, minimum: 0 }
                estimatedEffortPh: { type: number, minimum: 0, description: 'Admin only' }
                assigneeIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Task' }
        '400': { description: Sprint completed or parent task not in sprint }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get full task detail including sprint, steps, recent comments, recent attachments
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/TaskDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Tasks]
      summary: Update task fields (admin only)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:              { type: string }
                description:       { type: string }
                priority:          { $ref: '#/components/schemas/TaskPriority' }
                startDate:         { type: string, format: date }
                plannedDueDate:    { type: string, format: date }
                actualDueDate:     { type: string, format: date }
                plannedEffortPh:   { type: number, minimum: 0 }
                estimatedEffortPh: { type: number, minimum: 0 }
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Task' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Tasks]
      summary: Soft-delete a task and all sub-tasks (admin only)
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/status:
    patch:
      tags: [Tasks]
      summary: |
        Transition task status (assignee or admin).
        Valid: todo→in_progress, in_progress→review, review→done|in_progress, done→review.
        Completing a task sets `actualDueDate` if not already set.
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/TaskStatus' }
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Task' }
        '400': { description: Invalid status transition }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/effort:
    patch:
      tags: [Tasks]
      summary: Update effort hours (assignee or admin; staff cannot set estimatedEffortPh)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actualEffortPh:    { type: number, minimum: 0 }
                estimatedEffortPh: { type: number, minimum: 0, description: 'Admin only' }
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Task' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/assignees:
    get:
      tags: [Tasks]
      summary: Full assignee history (active and past)
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Assignee history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TaskAssigneeHistoryEntry' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Tasks]
      summary: |
        Add or remove assignees in a single call (admin only).
        Users in `add` that are already assigned are silently skipped.
        Users in `remove` that are not assigned are silently skipped.
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add:
                  type: array
                  items: { type: string, format: uuid }
                  description: User IDs to assign (must be active staff)
                remove:
                  type: array
                  items: { type: string, format: uuid }
                  description: User IDs to unassign
            examples:
              assign:
                summary: Assign two users
                value: { add: ['uuid-1', 'uuid-2'], remove: [] }
              reassign:
                summary: Swap assignee
                value: { add: ['uuid-new'], remove: ['uuid-old'] }
      responses:
        '200':
          description: Task with updated assignee list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Task' }
        '400': { description: One or more users in `add` are not active staff }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...